Refactor load_config method
Move load_config to config.py, remove duplicate declaration in mdns.py, update usage in test_comitup.py.
