+++ comitup/config.py
@@ -12,6 +12,15 @@
 
 import configparser
 import io
+import os
+import random
+import shutil
+
+from comitup import persist
+
+PERSIST_PATH = "/var/lib/comitup/comitup.json"
+CONF_PATH = "/etc/comitup.conf"
+BOOT_CONF_PATH = "/boot/comitup.conf"
 
 
 class Config(object):
@@ -32,3 +41,31 @@ class Config(object):
             return self._config.get(self._section, tag)
         except configparser.NoOptionError:
             raise AttributeError
+
+
+def load_data():
+    if os.path.isfile(BOOT_CONF_PATH):
+        try:
+            dest = shutil.copyfile(BOOT_CONF_PATH, CONF_PATH)
+            print("Boot config file copied:", dest)
+            os.remove(BOOT_CONF_PATH)
+        except:
+            print("Error occurred while copying file.")
+
+    conf = Config(
+                CONF_PATH,
+                defaults={
+                    'ap_name': 'comitup-<nnn>',
+                    'ap_password': '',
+                    'web_service': '',
+                    'service_name': 'workstation',
+                    'external_callback': '/usr/local/bin/comitup-callback',
+                },
+             )
+
+    data = persist.persist(
+                PERSIST_PATH,
+                {'id': str(random.randrange(1000, 9999))},
+           )
+
+    return (conf, data)
